+++ src/wifi/model/wifi-phy.h
@@ -650,7 +650,7 @@ class WifiPhy : public Object
      * \param staId the STA-ID
      */
     void NotifyMonitorSniffRx(Ptr<const WifiPsdu> psdu,
-                              uint16_t channelFreqMhz,
+                              double channelFreqMhz,
                               WifiTxVector txVector,
                               SignalNoiseDbm signalNoise,
                               std::vector<bool> statusPerMpdu,
@@ -698,7 +698,7 @@ class WifiPhy : public Object
      * \param staId the STA-ID
      */
     void NotifyMonitorSniffTx(Ptr<const WifiPsdu> psdu,
-                              uint16_t channelFreqMhz,
+                              double channelFreqMhz,
                               WifiTxVector txVector,
                               uint16_t staId = SU_STA_ID);
 
@@ -985,7 +985,7 @@ class WifiPhy : public Object
     /**
      * \return the operating center frequency (MHz)
      */
-    uint16_t GetFrequency() const;
+    double GetFrequency() const;
     /**
      * \return the index of the primary 20 MHz channel
      */
@@ -1272,7 +1272,7 @@ class WifiPhy : public Object
     /**
      * \return the subcarrier spacing corresponding to the configure standard (Hz)
      */
-    uint32_t GetSubcarrierSpacing() const;
+    double GetSubcarrierSpacing() const;
 
     /**
      * Callback invoked when the PHY model starts to transmit a signal
