+++ src/wifi/model/wifi-ppdu.h
@@ -140,7 +140,7 @@ class WifiPpdu : public SimpleRefCount<WifiPpdu>
     /**
      * \return the center frequency per segment (MHz) used for the transmission of this PPDU
      */
-    std::vector<uint16_t> GetTxCenterFreqs() const;
+    std::vector<double> GetTxCenterFreqs() const;
 
     /**
      * Check whether the given PPDU overlaps a given channel.
@@ -149,7 +149,7 @@ class WifiPpdu : public SimpleRefCount<WifiPpdu>
      * \param maxFreq the maximum frequency (MHz) of the channel
      * \return true if this PPDU overlaps the channel, false otherwise
      */
-    bool DoesOverlapChannel(uint16_t minFreq, uint16_t maxFreq) const;
+    bool DoesOverlapChannel(double minFreq, double maxFreq) const;
 
     /**
      * Get the modulation used for the PPDU.
@@ -200,12 +200,12 @@ class WifiPpdu : public SimpleRefCount<WifiPpdu>
      */
     virtual std::string PrintPayload() const;
 
-    WifiPreamble m_preamble;          //!< the PHY preamble
-    WifiModulationClass m_modulation; //!< the modulation used for the transmission of this PPDU
-    WifiConstPsduMap m_psdus;         //!< the PSDUs contained in this PPDU
-    std::vector<uint16_t> m_txCenterFreqs; //!< the center frequency (MHz) per segment used for the
-                                           //!< transmission of this PPDU
-    uint64_t m_uid;                        //!< the unique ID of this PPDU
+    WifiPreamble m_preamble;             //!< the PHY preamble
+    WifiModulationClass m_modulation;    //!< the modulation used for the transmission of this PPDU
+    WifiConstPsduMap m_psdus;            //!< the PSDUs contained in this PPDU
+    std::vector<double> m_txCenterFreqs; //!< the center frequency (MHz) per segment used for the
+                                         //!< transmission of this PPDU
+    uint64_t m_uid;                      //!< the unique ID of this PPDU
     mutable std::optional<WifiTxVector>
         m_txVector; //!< the TXVECTOR at TX PHY or the reconstructed TXVECTOR at RX PHY (or
                     //!< std::nullopt if TXVECTOR has not been reconstructed yet)
