+++ src/wifi/model/wifi-spectrum-phy-interface.h
@@ -99,7 +99,7 @@ class WifiSpectrumPhyInterface : public SpectrumPhy
      * \return the center frequency in MHz for each segment of the the spectrum channel this
      * interface is attached to to
      */
-    const std::vector<uint16_t>& GetCenterFrequencies() const;
+    const std::vector<double>& GetCenterFrequencies() const;
 
     /**
      * Get the channel width in MHz covered by the spectrum channel this interface is attached to
@@ -124,7 +124,7 @@ class WifiSpectrumPhyInterface : public SpectrumPhy
      * \param bandBandwidth the width of each band in Hz
      * \param guardBandwidth the total width of the guard band in MHz
      */
-    void SetRxSpectrumModel(const std::vector<uint16_t>& centerFrequencies,
+    void SetRxSpectrumModel(const std::vector<double>& centerFrequencies,
                             ChannelWidthMhz channelWidth,
                             uint32_t bandBandwidth,
                             ChannelWidthMhz guardBandwidth);
@@ -162,7 +162,7 @@ class WifiSpectrumPhyInterface : public SpectrumPhy
     Ptr<SpectrumWifiPhy> m_spectrumWifiPhy;     ///< spectrum PHY
     Ptr<NetDevice> m_netDevice;                 ///< the device
     Ptr<SpectrumChannel> m_channel;             ///< spectrum channel
-    std::vector<uint16_t> m_centerFrequencies;  ///< center frequency per segment in MHz
+    std::vector<double> m_centerFrequencies;    ///< center frequency per segment in MHz
     ChannelWidthMhz m_channelWidth;             ///< channel width in MHz
     Ptr<const SpectrumModel> m_rxSpectrumModel; ///< receive spectrum model
